services/core/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask == 2.2.2
Unidecode == 1.3.6
nltk == 3.7
python_dotenv == 0.21.0
requests == 2.28.1
waitress == 2.1.2
